pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "longshade"
version = "0.1.0"
description = "Conversable persona generation from personal data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Alex Towell", email = "lex@metafunctor.com" }
]
keywords = [
    "persona",
    "llm",
    "rag",
    "voice",
    "archiving",
    "echo",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "sentence-transformers>=2.0.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.19.0",
    "PyYAML>=5.4",
    "python-frontmatter>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.scripts]
longshade = "longshade.cli:main"

[project.urls]
Homepage = "https://github.com/aarontowell/longshade"
Documentation = "https://github.com/aarontowell/longshade#readme"
Repository = "https://github.com/aarontowell/longshade"

[tool.hatch.build.targets.wheel]
packages = ["src/longshade"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
